place order api

+++ jwt/src/main/java/com/act/ecommerce/dao/ProductDao.java
@@ -3,5 +3,5 @@ package com.act.ecommerce.dao;
 import com.act.ecommerce.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductDao extends JpaRepository<Product, Integer> {
+public interface ProductDao extends JpaRepository<Product, Long> {
 }
